added authorization for blog delete and tokenetractor middleware

+++ part4/blog-list/controllers/blogs.js
@@ -1,22 +1,37 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1})
     response.json(blogs)
 })
-  
+
+// const getTokenFrom = request => {
+//     const authorization = request.get('authorization')
+//     if (authorization && authorization.startsWith('Bearer ')) {
+//         console.log(authorization)
+//         return authorization.replace('Bearer ', '')
+//     }
+//     return null
+// }
+
 blogRouter.post('/', async (request, response) => {
     const body = request.body
-    const user = await User.findById(body.userId)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid'})
+    }
+
+    const user = await User.findById(decodedToken.id)
     
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
         likes: body.likes,
-        user: body.userId === undefined ? undefined : user._id
+        user: user._id
     })
 
     if (!blog.likes || blog.likes < 0) {
@@ -36,6 +51,18 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid'})
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() !== user.id.toString()) {
+        return response.status(401).json({ error: 'invalid id'})
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
